# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HTMLPush API",
    "description": "HTMLPush 是一个让手机浏览器变成桌面显示器的服务。后台上传或切换 HTML 内容，手机端读取同一份 API 文档产物。",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    }
  ],
...
  "components": {
    "schemas": {
      "PageList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PageInfo"
        }
      },
      "PageInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "页面唯一 ID"
          },
          "name": {
            "type": "string",
            "description": "页面名称"
          },
          "filename": {
            "type": "string",
            "description": "存储文件名"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "description": "上传时间"
          }
        },
        "required": [
          "id",
          "name",
          "filename",
          "uploadedAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误信息"
          }
        },
        "required": [
          "error"
        ]
      },
      "PageUploadRequest": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "HTML 文件"
          },
          "name": {
            "type": "string",
            "description": "页面名称，可选，默认使用文件名"
          }
        },
        "required": [
          "file"
        ]
      },
...
      "ImageList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ImageAsset"
        }
      },
      "ImageAsset": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ImageAssetBase"
          },
          {
            "type": "object",
            "properties": {
              "imageUrl": {
                "type": "string",
                "description": "原图访问地址"
              },
              "pageUrl": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "关联页面访问地址"
              }
            },
            "required": [
              "imageUrl",
              "pageUrl"
            ]
          }
        ]
      },
      "ImageAssetBase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "图片唯一 ID"
          },
          "name": {
            "type": "string",
            "description": "图片名称"
          },
          "filename": {
            "type": "string",
            "description": "存储文件名"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME 类型"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "description": "文件大小，单位字节"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "description": "上传时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          },
          "pageId": {
            "type": [
              "string",
              "null"
            ],
            "description": "关联页面 ID"
          }
        },
        "required": [
          "id",
          "name",
          "filename",
          "mimeType",
          "size",
          "uploadedAt",
          "updatedAt",
          "pageId"
        ]
      },
      "ImageSort": {
        "type": "string",
        "enum": [
          "uploadedAt-desc",
          "uploadedAt-asc"
        ],
        "description": "排序方式，默认按上传时间倒序"
      },
      "ImageUploadResponse": {
        "type": "object",
        "properties": {
          "image": {
            "$ref": "#/components/schemas/ImageAsset"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "image",
          "page"
        ]
      },
      "ImageUploadRequest": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "图片文件"
          },
          "name": {
            "type": "string",
            "description": "图片名称"
          }
        },
        "required": [
          "file"
        ]
      },
      "DeletedIdsResponse": {
        "type": "object",
        "properties": {
          "deletedIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "已删除图片 ID 列表"
          }
        },
        "required": [
          "deletedIds"
        ]
      },
      "ImageBatchDeleteRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1,
            "description": "待删除图片 ID 列表"
          }
        },
        "required": [
          "ids"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/pages": {
      "get": {
        "tags": [
          "页面管理"
        ],
        "summary": "获取所有页面列表",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "页面管理"
        ],
        "summary": "上传 HTML 文件",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PageUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageInfo"
                }
              }
            }
          },
          "400": {
            "description": "缺少文件",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "上传失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images": {
      "get": {
        "tags": [
          "图床管理"
        ],
        "summary": "获取图片列表",
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/ImageSort"
            },
            "required": false,
            "description": "排序方式，默认按上传时间倒序",
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "图床管理"
        ],
        "summary": "上传图片并生成展示页",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "缺少文件",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "文件过大",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "文件类型不支持",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "上传失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "图床管理"
        ],
        "summary": "批量删除图片",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageBatchDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedIdsResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "批量删除失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}